+++ exec/js.go
@@ -10,16 +10,19 @@ import (
 // `dir` is expected to be the path to a local directory
 // which contains the JS code to be executed. The container
 // will run the `main.js` file present in this directory.
-func RunJS(dir string) error {
-	cfg := config(dir)
+//
+// `main` is the path, local to `dir`, of the main JS file
+// to be run by the container.
+func RunJS(dir string, main string) error {
+	cfg := config(dir, main)
 	js.Run(cfg)
 	return nil
 }
 
-func config(dir string) js.DockerConfig {
+func config(dir string, main string) js.DockerConfig {
 	var cfg = js.DockerConfig{
 		Image:      "node:4",
-		Command:    []string{"node", "./test.js"},
+		Command:    []string{"node", main},
 		Volumes:    map[string]struct{}{"/usr/src/app": {}},
 		WorkingDir: "/usr/src/app",
 		Binds:      []string{dir + ":/usr/src/app"},