+++ exec/js/test.js
@@ -1,6 +1,0 @@
-var data = require("./data");
-
-console.log("Hello world from JS inside Docker!");
-console.log("Value from data entries: " + data.test);
-
-module.exports = [];